+++ compysition/testutils/__init__.py
@@ -1,0 +1,23 @@
+import random
+import sys
+from types import ModuleType, FunctionType
+from gc import get_referents
+
+def gen_rdm_attr_name(length):
+    letters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ_"
+    return "".join([random.choice(letters) for i in range(length)])
+
+def getsize(obj):
+    BLACKLIST = type, ModuleType, FunctionType
+    if isinstance(obj, BLACKLIST):
+        raise TypeError('getsize() does not take argument of type: '+ str(type(obj)))
+    seen_ids, size, objects = set(), 0, [obj]
+    while objects:
+        need_referents = []
+        for obj in objects:
+            if not isinstance(obj, BLACKLIST) and id(obj) not in seen_ids:
+                seen_ids.add(id(obj))
+                size += sys.getsizeof(obj)
+                need_referents.append(obj)
+        objects = get_referents(*need_referents)
+    return size